Move reusable build tasks to separated files
