+++ postcss.config.js
@@ -49,5 +49,5 @@ module.exports = {
     // Add vendor prefixes to CSS rules using values from caniuse.com
     // https://github.com/postcss/autoprefixer
     require('autoprefixer')(),
-  ];
+  ],
 };
